+++ panov.ilya/T1/Diamond.cpp
