+++ panov.ilya/T1/Diamond.h
