+++ panov.ilya/T1/FrameRectangle.cpp
