+++ panov.ilya/T1/FrameRectangle.h
