+++ panov.ilya/T1/Point.cpp
