+++ panov.ilya/T1/figure.cpp
@@ -1,4 +1,4 @@
-#include <iostream>
+﻿#include <iostream>
 #include <iomanip>
 #include "Shape.h"
 #include "Point.h"
